requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
pymongo==3.11
APScheduler==3.10.4
pyfiglet==1.0.2
rich==13.6.0
opencv-python==4.8.1.78
requests==2.31.0
PyNaCl==1.5.0
torch==2.1.0
torchvision==0.16.0
transformers==4.34.0
